cleanup(misc): clean up init generators
