+++ packages/jest/src/generators/configuration/lib/update-vscode-recommended-extensions.ts
@@ -1,0 +1,16 @@
+import { updateJson, type Tree } from '@nx/devkit';
+
+export function updateVsCodeRecommendedExtensions(host: Tree) {
+  if (!host.exists('.vscode/extensions.json')) {
+    return;
+  }
+
+  updateJson(host, '.vscode/extensions.json', (json) => {
+    json.recommendations = json.recommendations || [];
+    const extension = 'firsttris.vscode-jest-runner';
+    if (!json.recommendations.includes(extension)) {
+      json.recommendations.push(extension);
+    }
+    return json;
+  });
+}
