+++ packages/rollup/src/generators/init/init.spec.ts
@@ -1,4 +1,4 @@
-import { Tree, readJson, NxJsonConfiguration, updateJson } from '@nx/devkit';
+import { Tree, readJson } from '@nx/devkit';
 import { createTreeWithEmptyWorkspace } from '@nx/devkit/testing';
 import { nxVersion } from '../../utils/versions';
 
@@ -11,33 +11,14 @@ describe('rollupInitGenerator', () => {
     tree = createTreeWithEmptyWorkspace({ layout: 'apps-libs' });
   });
 
-  it('should support swc', async () => {
-    await rollupInitGenerator(tree, { compiler: 'swc' });
+  it('should install deps', async () => {
+    await rollupInitGenerator(tree, {});
 
     const packageJson = readJson(tree, 'package.json');
     expect(packageJson).toEqual({
       name: expect.any(String),
       dependencies: {},
-      devDependencies: {
-        '@nx/rollup': nxVersion,
-        '@swc/helpers': expect.any(String),
-        '@swc/core': expect.any(String),
-        'swc-loader': expect.any(String),
-      },
-    });
-  });
-
-  it('should support tsc', async () => {
-    await rollupInitGenerator(tree, { compiler: 'tsc' });
-
-    const packageJson = readJson(tree, 'package.json');
-    expect(packageJson).toEqual({
-      name: expect.any(String),
-      dependencies: {},
-      devDependencies: {
-        '@nx/rollup': nxVersion,
-        tslib: expect.any(String),
-      },
+      devDependencies: { '@nx/rollup': nxVersion },
     });
   });
 });
