+++ packages/rollup/src/generators/init/init.ts
@@ -4,39 +4,14 @@ import {
   GeneratorCallback,
   Tree,
 } from '@nx/devkit';
+import { nxVersion } from '../../utils/versions';
 import { Schema } from './schema';
-import { swcCoreVersion, swcHelpersVersion } from '@nx/js/src/utils/versions';
-import {
-  nxVersion,
-  swcLoaderVersion,
-  tsLibVersion,
-} from '../../utils/versions';
 
 export async function rollupInitGenerator(tree: Tree, schema: Schema) {
   let task: GeneratorCallback = () => {};
 
   if (!schema.skipPackageJson) {
-    if (schema.compiler === 'swc') {
-      task = addDependenciesToPackageJson(
-        tree,
-        {},
-        {
-          '@nx/rollup': nxVersion,
-          '@swc/helpers': swcHelpersVersion,
-          '@swc/core': swcCoreVersion,
-          'swc-loader': swcLoaderVersion,
-        }
-      );
-    } else {
-      task = addDependenciesToPackageJson(
-        tree,
-        {},
-        {
-          '@nx/rollup': nxVersion,
-          tslib: tsLibVersion,
-        }
-      );
-    }
+    task = addDependenciesToPackageJson(tree, {}, { '@nx/rollup': nxVersion });
   }
 
   if (!schema.skipFormat) {
