+++ packages/rollup/src/generators/init/schema.d.ts
@@ -1,5 +1,4 @@
 export interface Schema {
-  compiler?: 'babel' | 'swc' | 'tsc';
   skipFormat?: boolean;
   skipPackageJson?: boolean;
 }
