+++ packages/rollup/src/generators/init/schema.json
@@ -6,12 +6,6 @@
   "description": "Init Webpack Plugin.",
   "type": "object",
   "properties": {
-    "compiler": {
-      "type": "string",
-      "enum": ["babel", "swc", "tsc"],
-      "description": "The compiler to initialize for.",
-      "default": "babel"
-    },
     "skipFormat": {
       "description": "Skip formatting files.",
       "type": "boolean",
