+++ packages/rollup/src/utils/ensure-dependencies.spec.ts
@@ -1,0 +1,37 @@
+import { readJson, type Tree } from '@nx/devkit';
+import { createTreeWithEmptyWorkspace } from '@nx/devkit/testing';
+import { ensureDependencies } from './ensure-dependencies';
+
+describe('ensureDependencies', () => {
+  let tree: Tree;
+
+  beforeEach(() => {
+    tree = createTreeWithEmptyWorkspace({ layout: 'apps-libs' });
+  });
+
+  it('should support swc', () => {
+    ensureDependencies(tree, { compiler: 'swc' });
+
+    const packageJson = readJson(tree, 'package.json');
+    expect(packageJson).toEqual({
+      name: expect.any(String),
+      dependencies: {},
+      devDependencies: {
+        '@swc/helpers': expect.any(String),
+        '@swc/core': expect.any(String),
+        'swc-loader': expect.any(String),
+      },
+    });
+  });
+
+  it('should support tsc', () => {
+    ensureDependencies(tree, { compiler: 'tsc' });
+
+    const packageJson = readJson(tree, 'package.json');
+    expect(packageJson).toEqual({
+      name: expect.any(String),
+      dependencies: {},
+      devDependencies: { tslib: expect.any(String) },
+    });
+  });
+});
