+++ packages/storybook/src/generators/configuration/configuration.ts
@@ -45,6 +45,8 @@ import {
   tsNodeVersion,
 } from '../../utils/versions';
 import { interactionTestsDependencies } from './lib/interaction-testing.utils';
+import { ensureDependencies } from './lib/ensure-dependencies';
+import { editRootTsConfig } from './lib/edit-root-tsconfig';
 
 export async function configurationGenerator(
   tree: Tree,
@@ -93,11 +95,11 @@ export async function configurationGenerator(
     skipFormat: true,
   });
   tasks.push(jsInitTask);
-  const initTask = await initGenerator(tree, {
-    uiFramework: schema.uiFramework,
-    skipFormat: true,
-  });
+  const initTask = await initGenerator(tree, { skipFormat: true });
   tasks.push(initTask);
+  tasks.push(ensureDependencies(tree, { uiFramework: schema.uiFramework }));
+
+  editRootTsConfig(tree);
 
   const nxJson = readNxJson(tree);
   const hasPlugin = nxJson.plugins?.some((p) =>
@@ -111,7 +113,8 @@ export async function configurationGenerator(
       ? 'components'
       : 'src';
 
-  const usesVite = !!viteConfigFilePath || schema.uiFramework.endsWith('-vite');
+  const usesVite =
+    !!viteConfigFilePath || schema.uiFramework?.endsWith('-vite');
 
   createProjectStorybookDir(
     tree,
@@ -216,7 +219,7 @@ export async function configurationGenerator(
     devDeps['core-js'] = coreJsVersion;
   }
 
-  if (schema.uiFramework.endsWith('-vite') && !viteConfigFilePath) {
+  if (schema.uiFramework?.endsWith('-vite') && !viteConfigFilePath) {
     // This means that the user has selected a Vite framework
     // but the project does not have Vite configuration.
     // We need to install the @nx/vite plugin in order to be able to use
