+++ packages/storybook/src/generators/init/schema.d.ts
@@ -1,7 +1,4 @@
-import { UiFramework } from '../../utils/models';
-
 export interface Schema {
-  uiFramework: UiFramework;
   skipFormat?: boolean;
   skipPackageJson?: boolean;
 }
